Make Dockerization in CloudFocker possible
by creating a stub Procfile.

The following line in the buildpack lifecycle means that you must have a
Procfile with a web command in order to successfully run a buildpack:

https://github.com/cloudfoundry-incubator/buildpack_app_lifecycle/blob/master/buildpackrunner/runner.go#L95

You can get past this check by providing a dummy command in the
Procfile, as is done in this commit.

[finishes #89807954]